billsOCR/model/requirements.txt
torch
torchvision>=0.5
opencv-python-headless
Pillow
scikit-image
python-bidi
PyYAML
Shapely
ninja
editdistance
Polygon3
pyclipper
Cython
prefetch_generator
scipy
yacs
tqdm
opencv-python==4.6.0.66
mlflow
boto3
psycopg2-binary
python-dotenv
pynvml